Move header helpers and serializer into separate folder
